aws exception handling

+++ grails-app/controllers/web106/file/upload/UploadFileController.groovy
@@ -24,72 +24,7 @@ class UploadFileController {
      */
 	def uploadImage() {
 
-		def activeWebsite
-		
-		def activeWebsiteSession = session.getAttribute('activeWebsite')
-		activeWebsite = activeWebsiteSession
-		Website website = Website.find {
-			id == activeWebsite
-		}
-
-		def multiPartFile = request.getFile('file')
-		if(multiPartFile.empty) {
-			flash.message = "File cannot be empty"
-		} else {
-			File convertedFile = fileTypeConverterService.convert(multiPartFile)
-			if(convertedFile == null){
-				flash.message = "File must be .jpg or .png"
-			} else {
-
-                def prefix =  website.workGroup.name+ "/" + website.title + "/" + ResourceHolder.bucketprefixForImages + "/"
-
-                //upload to bucket/workgroup/site/imagebucket
-                uploadS3Service.uploadFileToS3Bucket(ResourceHolder.bucketGlobal,convertedFile, prefix)
-            }
-
-		}
-
-
-
-
-		redirect (action:'index')
-	}
-
-    def list() {
-
-        def activeWebsite
-
-        def activeWebsiteSession = session.getAttribute('activeWebsite')
-        activeWebsite = activeWebsiteSession
-        Website website = Website.find {
-            id == activeWebsite
-        }
-
-        def prefix =  website.workGroup.name+ "/" + website.title + "/" + ResourceHolder.bucketprefixForImages + "/"
-        def list = uploadS3Service.FileNamesInBucket(ResourceHolder.bucketGlobal, prefix)
-
-        def pictures= []
-
-        list.each {
-            def item = [:]
-            item['name'] = it
-            item['url'] ='http://s3-eu-west-1.amazonaws.com/'+ ResourceHolder.bucketGlobal + '/' + prefix + it
-            item['fileNameID'] = it.replace('.', '_')
-
-            pictures.add(item)
-        }
-
-
-
-        def model = [:]
-        model['pictures'] = pictures
-        render view : 'list' , model : model
-    }
-
-    def delete() {
-
-        if(params.id) {
-
+        try {
 
             def activeWebsite
 
@@ -99,14 +34,103 @@ class UploadFileController {
                 id == activeWebsite
             }
 
-            String keyFileName = params.id
-            int last = keyFileName.lastIndexOf('_')
-            String keyFileNameNew = new StringBuilder(keyFileName).replace(last, last+1,".").toString();
+            def multiPartFile = request.getFile('file')
+            if(multiPartFile.empty) {
+                flash.message = "File cannot be empty"
+            } else {
+                File convertedFile = fileTypeConverterService.convert(multiPartFile)
+                if(convertedFile == null){
+                    flash.message = "File must be .jpg or .png"
+                } else {
+
+                    def prefix =  website.workGroup.name+ "/" + website.title + "/" + ResourceHolder.bucketprefixForImages + "/"
+
+                    //upload to bucket/workgroup/site/imagebucket
+                    uploadS3Service.uploadFileToS3Bucket(ResourceHolder.bucketGlobal,convertedFile, prefix)
+                }
+
+            }
+
+		    redirect (action:'index')
+        } catch (AmazonServiceException) {
+            redirect controller: 'errorsWeb106' ,view: 'aws'
+        } catch (AmazonClientException) {
+            redirect controller: 'errorsWeb106' ,view: 'aws'
+        } catch (UnexpectedRollbackException) {
+            redirect controller: 'errorsWeb106' ,view: 'aws'
+        }
+	}
+
+    def list() {
+
+        try {
+
+            def activeWebsite
+
+            def activeWebsiteSession = session.getAttribute('activeWebsite')
+            activeWebsite = activeWebsiteSession
+            Website website = Website.find {
+                id == activeWebsite
+            }
 
             def prefix =  website.workGroup.name+ "/" + website.title + "/" + ResourceHolder.bucketprefixForImages + "/"
-            uploadS3Service.deleteFile(ResourceHolder.bucketGlobal,prefix, keyFileNameNew )
+            def list = uploadS3Service.FileNamesInBucket(ResourceHolder.bucketGlobal, prefix)
+
+            def pictures= []
+
+            list.each {
+                def item = [:]
+                item['name'] = it
+                item['url'] ='http://s3-eu-west-1.amazonaws.com/'+ ResourceHolder.bucketGlobal + '/' + prefix + it
+                item['fileNameID'] = it.replace('.', '_')
+
+                pictures.add(item)
+            }
+
+
+
+            def model = [:]
+            model['pictures'] = pictures
+            render view : 'list' , model : model
+        } catch (AmazonServiceException) {
+            redirect controller: 'errorsWeb106' ,view: 'aws'
+        } catch (AmazonClientException) {
+            redirect controller: 'errorsWeb106' ,view: 'aws'
+        } catch (UnexpectedRollbackException) {
+            redirect controller: 'errorsWeb106' ,view: 'aws'
+        }
+    }
+
+    def delete() {
+
+        try {
+
+            if(params.id) {
+
+                def activeWebsite
+
+                def activeWebsiteSession = session.getAttribute('activeWebsite')
+                activeWebsite = activeWebsiteSession
+                Website website = Website.find {
+                    id == activeWebsite
+                }
+
+                String keyFileName = params.id
+                int last = keyFileName.lastIndexOf('_')
+                String keyFileNameNew = new StringBuilder(keyFileName).replace(last, last+1,".").toString();
+
+                def prefix =  website.workGroup.name+ "/" + website.title + "/" + ResourceHolder.bucketprefixForImages + "/"
+                uploadS3Service.deleteFile(ResourceHolder.bucketGlobal,prefix, keyFileNameNew )
+            }
+            redirect action: 'list'
+
+        } catch (AmazonServiceException) {
+            redirect controller: 'errorsWeb106' ,view: 'aws'
+        } catch (AmazonClientException) {
+            redirect controller: 'errorsWeb106' ,view: 'aws'
+        } catch (UnexpectedRollbackException) {
+            redirect controller: 'errorsWeb106' ,view: 'aws'
         }
 
-        redirect action: 'list'
     }
 }
